+++ pages/login.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import LoginForm from '../components/login/LoginForm';
-import LoginHeader from '../components/Layout/LoginHeader';
+import LoginHeader from '../components/header/LoginHeader';
 import SignupButton from '../components/login/SignupButton';
 
 const LoginPage = () => {
